requirements.txt
elasticsearch>=8.0.0
PyPDF2>=3.0.0
pyyaml>=6.0
requests>=2.28.0
python-dateutil>=2.8.0
